Added contact info
I'm full of myself lol
